Begin adding control for multiple exposures
Photo kicks off exposures.
The main loop checks for how many are remaining and set time for next
exposure.
This is somewhat working but reporting excessively and erroneously in
the main loop the "Exposing now at:" message while counting down the
exposures. It is close enough that I though I should commit it.